2022-02-09 15:07 kdg

+++ oneTouch/src/main/java/com/onetouch/web/mtr/in/service/impl/MtrInServiceImpl.java
@@ -51,12 +51,14 @@ public class MtrInServiceImpl implements MtrInService {
 				if(vo.getFltAmt() == null || vo.getFltAmt() == "") {
 					vo.setFltAmt("0");
 				}
-				System.out.println("notInAmt");
-				System.out.println(vo.getNotinAmt());
-				System.out.println("fltAmt");
-				System.out.println(vo.getFltAmt());
-				inMapper.insertIn(vo);
-				inMapper.plusOrd(vo);
+				if( vo.getInAmt() == "0") {
+					System.out.println("발주업데이트만");
+					inMapper.plusOrd(vo);
+				} else {
+					System.out.println("입고insert+발주update");
+					inMapper.insertIn(vo);
+					inMapper.plusOrd(vo);
+				}
 				
 				inAmt = Integer.parseInt(vo.getInAmt());
 				mngAmt = Integer.parseInt(vo.getMngAmt());
@@ -80,7 +82,9 @@ public class MtrInServiceImpl implements MtrInService {
 		}
 		if(mvo.getUpdatedRows() != null) {
 		    for(MtrInVO data : mvo.getUpdatedRows()){
-		    	inMapper.updateIn(data);
+			    	inMapper.plusOrd(data);
+			    	inMapper.minusOrd(data);
+			    	inMapper.updateIn(data);
 		    	};
 		}
 	}
@@ -88,7 +92,9 @@ public class MtrInServiceImpl implements MtrInService {
 	@Override
 	public void deleteIn(ModifyVO<MtrInVO> mvo) {
 		for(MtrInVO data : mvo.getDeletedRows()){
-	    	inMapper.deleteIn(data);
+			if(data.getInNo() != null) {
+				inMapper.deleteIn(data);
+			}
 	    	inMapper.minusOrd(data);
 		};
 	}
